Implementing overlays
The general idea here is to have a way to call, from the `ComparisonProxy`:

```
comparison.get_overlay(OverlayType.line_execution_count)
```

which takes care of loading the whole data from storage as needed (and can cache some stuff), and then use both the comparison and the loaded data for all the logic it needs.

This is not ready. It still requires the `_load_critical_path_report` to be implemented, along with the class that can actually use the data from the summarization (loaded from storage) and implement things like `get_critical_files_filenames`

But this is a general view on how the system can work. It also helps us prepare the ground for possible "future overlays"