[Step 1] Adding setex to RefreshService
Adding `setex` to RefreshService. This way we have a expirable key that will not block customers from calling refresh.

But for now we need to keep both until at least a second deploy, so we don't have race conditions with customers that started the refreshing without the setex and then have something break in case setex creates some unexpected expectation.