Enforcing secrets to be used on same repository
A long time ago, yaml secrets ( https://docs.codecov.com/docs/codecov-yaml#how-do-i-encrypt-data-that-i-do-not-want-public ) could be created on a specific repo page. And we enforced that repos created on a given repo were used only on that repo.

During the porting, we accidentally ended up relaxing that requirement.

Here is how a malicious party can abuse this:
1) Find a public repo using codecov secrets in their codecov.yml
2) Copy the secrets to your own codecov.yml
3) Spin a codecov enterprise instance
4) Make upload from that repository to the enterprise instance
5) Watch the logs show the decrypted value

Here are some reasons why the attack is not easily doable/less valuable:
1) The attacker needs previous access to the secret in the codecov.yml, which usually means a public repository
2) The attacker needs to be willing to spin an enterprise instance( we don't show decrypted secrets in the UI)
3) The attacker needs to know that the secret leaks into the logs after decrypted
4) Very few customers use this feature
5) From looking at usage, most times this feature encrypts Slack channel webhooks (which do not allow reading of anything), which are not as valuable except for sending unauthorized messages, and is immediately noticed by the channel users.

The above reasons make it such attack not as valuable. But we should still at least implement this the way the feature was designed.

To be clear, an enterprise attacker that knows how this works can still manipulate the values in the db to make their repo look like the right values fro trigger a decryption of the secret. So this is still not entirely secure