Splitting install validation and useryaml validation
We now want to make it possible for the `secret` feature in user yaml ( https://docs.codecov.com/docs/codecov-yaml#how-do-i-encrypt-data-that-i-do-not-want-public  ) to use values inputted from the user on config.

For that, now the user yaml validation has to rely on config. But since config already relies on install validation, if install validation and user yaml validation were the same we would have some circular dependency (which can usually be circumvented with weird imports, but it's a bad practice).

So we are moving them to different places

This also helps in that `secret` feature is not meant to be used in the install yaml anyway, so making sure the decrypting doesn't happen there is a good step.

We are still not sure on how we will implement this "install-defined yaml secret", but this move seems sane regardless.