Adding ordering to comparedto query
Basically the base picking code had a bug. We had a query for picking the "closest to base" alternative, which ordered all valid results by descending timestamp and picked the most recent one (which was the first result).

```
SELECT *
FROM commits
WHERE commits.repoid = 6711778
  AND commits.branch = 'master'
  AND (commits.pullid IS NULL
       OR commits.merged)
  AND commits.timestamp < '2021-06-25'::TIMESTAMP
  AND commits.ci_passed
ORDER BY commits.TIMESTAMP DESC LIMIT 1
```

But the part of the query that was doing ordering was not properly applied, so the ordering did not happen, and therefore the query was not requesting the most recent commit anymore.

The bug itself is simple, a one line change and adding some tests.

But the most interesting part is that the bug did not affect anything because of postgres query planning.
Since we didn't apply ordering to a SQL query and requested its first result, postgres would simply return the first result that naturally appeared on its searches.
And for whoever knows the query planner, what it basically does is, using some stats about the database, makes a plan on what it thinks is the quickest way of getting to a result. Among those stats are indexes size, and expected/approximate number of rows that each filter gives.
Our query without the ordering is:

```
SELECT *
  FROM commits
  WHERE commits.repoid = 6711778
    AND commits.branch = 'master'
    AND (commits.pullid IS NULL
         OR commits.merged)
    AND commits.timestamp < '2021-06-25'::timestamp
    AND commits.ci_passed;
```

It just happened that postgres, looking at the query, always decided to use this index to do the search:

```
create index commits_repoid_timestamp_desc on commits (repoid, timestamp desc);
```

Notice that, by using an index that has timestamp desc, postgres would always naturally start by looking at the highest possible timestamps and then started going down until it found a match for the other filters.

So, even though we forgot to put the "timestamp descending" ordering, postgres would naturally do it for us, so the bug never really happened.

That, until now.

As we said, postgres uses what it knows about its tables and indexes to make its decision about how to execute a query. The fact that it chose an index that helped us was just luck.
On June 27th, something in our commits table passed a certain threshold. We don't know exactly what the threshold was. It was possibly just number of rows and size of table. But it basically made postgres decide that using the above index was not worth it, so it switched to a different strategy. And the new strategy did not naturally followed the timestamp DESC order anymore.
Look at both EXPLAIN results:

```
thiago-ro@34:postgres> EXPLAIN SELECT *
  FROM commits
  WHERE commits.repoid = 6711778
    AND commits.branch = 'master'
    AND (commits.pullid IS NULL
         OR commits.merged)
    AND commits.timestamp < '2021-06-26'::timestamp
    AND commits.ci_passed;
+-----------------------------------------------------------------------------------------------------------+
| QUERY PLAN                                                                                                |
|-----------------------------------------------------------------------------------------------------------|
| Index Scan using commits_repoid_timestamp_desc on commits  (cost=0.56..9462.56 rows=394 width=725)        |
|   Index Cond: ((repoid = 6711778) AND ("timestamp" < '2021-06-26 00:00:00'::timestamp without time zone)) |
|   Filter: (((pullid IS NULL) OR merged) AND ci_passed AND (branch = 'master'::text))                      |
+-----------------------------------------------------------------------------------------------------------+
```

and

```
thiago-ro@34:postgres> EXPLAIN SELECT *
  FROM commits
  WHERE commits.repoid = 6711778
    AND commits.branch = 'master'
    AND (commits.pullid IS NULL
         OR commits.merged)
    AND commits.timestamp < '2021-06-27'::timestamp
    AND commits.ci_passed;
+-----------------------------------------------------------------------------------------------------------------+
| QUERY PLAN                                                                                                      |
|-----------------------------------------------------------------------------------------------------------------|
| Bitmap Heap Scan on commits  (cost=167.87..9376.79 rows=395 width=725)                                          |
|   Recheck Cond: ((repoid = 6711778) AND ("timestamp" < '2021-06-27 00:00:00'::timestamp without time zone))     |
|   Filter: (((pullid IS NULL) OR merged) AND ci_passed AND (branch = 'master'::text))                            |
|   ->  Bitmap Index Scan on commits_repoid_timestamp_desc  (cost=0.00..167.78 rows=2321 width=0)                 |
|         Index Cond: ((repoid = 6711778) AND ("timestamp" < '2021-06-27 00:00:00'::timestamp without time zone)) |
+-----------------------------------------------------------------------------------------------------------------+
```

Notice how the 2021-06-26 (and all queries with earlier dates) result uses an "Index Scan" strategy, while the 2021-06-27 (and all queries with later dates) use a "Bitmap Heap Scan" strategy

An explanation about the difference between both is https://stackoverflow.com/a/6592963 . While the "Bitmap heap scan" also uses the index for its benefit, it does not follow the index order. So its results won't be naturally ordered.

Checking the above with different repoids yield the same result. So this strategy has probably a more global threshold that we hit recently.
Because of this, we had this sudden increase in problems. All pull requests with bases that were created after 2021-06-27 started showing bad compared_to results. We didn't change anything to cause it. It just happened that the database hit a threshold where it changed the strategy and our bug started to actually affect the final result

So the fix is simply ensuring the ordering will be explicit on the query we send