Testing ignoring test files
Adding a codecov.yl to ignore

```
"**tests**/test_*.py"
```

files.

The rationale is that we don't want to punish or reward people from writing long or short tests. The code that we care whether it is being tested or not is the code that actually runs in production

Worker already does this. I want to see if we can consider this in the API as well.

This will likely result in a drop in coverage. But I believe this is beneficial and the new value is more representative of our coverage, and will help us more easily determine what code is actually lacking coverage.