Creating `ProcessingError` dataclass
Creating a `ProcessingError` dataclass which should contain all the information that is needed to store later for the upload.

We could have done this as an exception to be extended. But I feel the `ProcessingResult` containing every possible piece of data, successful or not, is a good way to center all the data in one place, specially if we are later added warning-like objects and more fields as well.

We are also adding a `is_retryable`. The idea here is that, when an `ProcessingError` has `is_retryable = True`, there is reasonable belief that a retry in a few seconds might solve the issue. One example for this would be that the uploaded file is missing from storage, meaning the user might still finish uploading it