Creating Simple Legacy Token Authentication
This change is meant to be a working authentication system to allow for repository-only actions.

The core pieces of this change are:

1. `RepositoryAsUser` is a class that is meant to behave as a user, but still it has all the data from the repository itself. We are not doing anything new with it so later the other authentication models can still play well with the views that use this authenticator. But still, having a separate class ensures this authentication is not accidentally used on other views without the proper care.
2. `RepositoryAuth` is the class that will contain all the permission-relevant data. It knows what permissions the user has, and what repo do those permissions belong to. This allows us for later expanding this for a more general scope-based permissions without compromising the current implementation
3. RepositoryLegacyTokenAuthentication is a simple authentication based on the `repository.upload_token` object. It will unblock our other API endpoints development.

Its idea is to not create any expectations on the `request.user` object. This way, later we can add "normal" user authentication to the views and still make it work with minimum changes.

I don't love it, but I think this is solid enough, and still flexible enough for us to build on top of