Reducing retries on upload error
Currently, whenever there is a really unexpected error (`except Exception`) inside the upload processing context, we simply do a retry in the hopes that this will work later. This is a behavior we inherited from legacy processor.

But looking at the log lines from `"Unable to process report %s"`, it seems that this is very unlikely to happen. Most tasks just keep retrying until the retry limit is hit.

So retrying is a bit useless, and consumes cpu time and "log attention" and sentry events from the rest of the system.

We are not totally removing the retries yet. All we are doing is reducing the number of retries from 5 to 1 for now.