Raising ReportEmptyError on individual upload contexts
In the past we had this flow:

```
previous_report = get_previous_report()
res = process_recent_upload(this_upload)
new_report = merge_reports(previous_report, res)
if new_report.is_empty():
    raise ReportEmptyError()
```

which is fine, but whenever `previous_report` is not empty, an exception will not be raised regardless of the new upload being empty or not. Meaning that if the customer makes a few good uploads, then a bad (empty) one, that one is not really noticed.

Since we are working on bubbling the errors to the user, this is something we need to change

So the new logic will be:

```
previous_report = get_previous_report()
res = process_recent_upload(this_upload)
if res.is_empty():
    raise ReportEmptyError()
new_report = merge_reports(previous_report, res)

```

This will make errors show up more often, but we believe there is no impact to any flows outside of this case