Changing save method
Changing save method on `validate_upload` in the hopes that a smaller update query might make it easier on some index locks, and on some triggers that might run on repo updates.

We don't know if that is impactful at all. But the function only seems to update those fours fields, so it's worth a try.

Also adding one line of logging.