Generating changes from rust comparison
Implementing a function to generate changes from comparisons

The main goal is to try to have a way to verify if our rust calculations are correct

For this, we are going to run `get_changes_using_rust` besides `get_changes` function on the worker and compare results.

We are not using `change.totals` for now. For now we will only see if the generated changes contain the same path and other fields. If all worker, we will consider possibly comparing change.totals. But the change.totals is not actively used on the worker and the API will not use that field directly anyway. And we might choose to just implement the change.totals.coverage calculation on the rust level directly, since it has a bunch of small complexities.