Replacing asyncio.run with async_to_sync
It looks like that Django preferred way of dealing with async-sync conversions is to use `async_to_sync`.

And recently we had a problem where some code simply stopped mid-execution nearby a asyncio.run call. There is no way to tell whether the asyncio.run was actually the problem. But since Django says:

> async_to_sync() is essentially a more powerful version of the asyncio.run() function in Python’s standard library. As well as ensuring threadlocals work, it also enables the thread_sensitive mode of sync_to_async() when that wrapper is used below it.

it looks like this is an objectively good change with no major costs, so we might as well try anyway