Moving `build_payload` to point where it is really used
In a similar idea as the one in https://github.com/codecov/worker/pull/581/files , we need to only call `build_payload` if we are actually sending notifications (which we don't do in cases of `flag_coverage_not_uploaded_behavior == "exclude"`

The above PR was done on status notifiers. This is on checks notifiers. Ideally, we would refactor that logic so both notifiers would use the same idea. But this might take a bit

Also adding a `None` check. On production this doesn't happen, but if it does, it breaks a few things. So adding a None check so, if a None happens, the "default behavior" (include) happens