Moving `get_filtered_comparison` for statuses as well
On the same way we did changed https://github.com/codecov/worker/pull/722 , we want to change the logic in statuses too so the `get_filtered_comparison` does not get called unnecessarily, because on big reports, that is some loading time that impacts the timeout rate.

The change on the checks had a good impact. So hopefully this one does as well

Again, ideally we should merge this and the one for checks, but small differences (like the location in `payload` where message gets set) make them different enough for an actual joining take a while