Remove usage of `using_integration` on SyncTeams
On `SyncTeamsTask`, there is currently a value called `using_integration`. That value determines whether we use the customer oauth or the Github App Integration to make the API calls.

This value comes from the time where there was a task called `Refresh` in legacy. Refresh basically did:

```
if using_integration:
    do_some_repo_syncing_on_integration()
else:
    if teams:
        do_team_syncing()
    if repos:
        do_repo_syncing()
```

And when it was ported to legacy, the `do_team_syncing()` and `do_repo_syncing()` were split each on its tasks: SyncTeamsTask and SyncReposTask. But then the `do_some_repo_syncing_on_integration()` was just put inside the `SyncReposTask` under a `using_integration` option.

`SyncTeamsTask` should not have used a `using_integration`, let alone one that actually changed its behavior.

Currently, calling it with `using_integration=True` causes the following error:

```
{"message":"Resource not accessible by integration","documentation_url":"https://docs.github.com/rest/reference/orgs#list-organization-memberships-for-the-authenticated-user"}
```

which seems to be happening for 6.5% of the users.

This change will only remove the `using_integration` option from SyncTeamsTask. IT won't break anything because `**kwargs` will hold the value if the API still passes it.

A different change should take care of splitting `using_integration` from the SyncReposTask. Because as the above flow showed, they are really different flows.