Making exception classes pickle better
Our Torngit exceptions currently contain the response object.

And it looks like the `Response` class from the `httpx` lib is not really picklable.

So when celery sends the Torngit exceptions from one process to another, it looks like it crashes.

And we don't really need the actual response object to be picklable. We don't use the whole thing. So we are just using the text of the response, which is picklable.