Adding Lock around SyncReposTask
On the API we already have mechanisms to avoid two SyncRepos from being triggered at the same time.

But other triggers, such as webhooks, can't benefit from that, because the `RefreshService` is designed around users triggering it.

So we are adding a lock on the worker side. The lock benefits both `using_integration` and `not using_integration` cases, but is a separate lock for both.

The reason for this change is a recent Github Webhook that triggered the SyncReposTask 530 times for the same user in a 3 hour window, including many times in the same minute. It caused many RateLimits from being hit and a lot of SoftTimeLimit (possibly from database locks)