Limiting the number of profiling uploads that are joined at each time
We found that joining all profiling uploads regardless of how many are there might lead to a timeout.

So we are limiting it to the first 500 at a time.

We are also increasing frequency of the command and decreasing the number of hours to wait, so we get quicker results