Using dot star as the default list of paths for a given flag
We have this option where customers can set which paths they want a flag to be limited to.

On another point they can set which flags they want on a carryforward/status/checks situation. This works like a filter: "I want to carryforward these flags" or "I want to show a status only involving the data related to those flags"

When the customer does that, we take the paths that are connected to this flag and use them as extra filters. The idea is that if a user want an `apple` filter and the apple flag is connected to `tree/.*` file only, we apply an extra path filter to the above filters as well. Most of the times those end up being a no-op, because the processing of the report already filtered out the wrong paths from the upload connected to that flag, and just by using a flag filter, we will get the same result. This mostly affects customers when they want to change the flag paths after the upload processing happens.

But when the user does not associate a flag with paths, this means they want the flag to be related to all paths. So it's more like an implicit `.*` path filter. When this is the only flag involved in the filter, it doesn't matter because a `.*` path filter is just like not having a path filter. So we just return a `[]` instead of `[.*]`

But when there are two flags, and one has explicit paths associated with it and the other doesn't, this is a problem because we only start using the paths from the first one, since the second one contributed with `[]` paths.

This didn't tend to happen before because most customers that set flags tend to either set paths on all flags or on none. And even those that didn't fall into either of those groups didn't tend to mix those two flags in any status or CFF situation.

This change aims to fix that. Now, if one flag in the `get_paths_from_flags` list does not have an explicit `paths` set, that means that flag won't be limited to any path, and therefore that flags list will also not apply any extra path filter.

Alternatively we could make it so, when a customer set paths in flags in yaml, we could enforce that they set explicit paths on all flags. This way there would be a less implicit thing.

Or we could also remove this `get_paths_from_flags`, since as we said above, this is almost always a no-op, and the times where it is not a no-op, we are not sure whether it actually should happen